got rid of the HomeContainer file as it seemed a little redundant
