+++ src/App.jsx
@@ -2,7 +2,6 @@
 import { RouterProvider, createBrowserRouter } from "react-router-dom";
 import PersonalPage from "./components/PersonalPage/PersonalPage";
 import Home from "./components/Home/Home";
-import HomeContainer from "./container/HomeContainer";
 
 function App() {
   const appRoutes = createBrowserRouter([
@@ -11,7 +10,7 @@ function App() {
       children: [
         {
           path: "/",
-          element: <HomeContainer />,
+          element: <Home />,
         },
         {
           path: "/:userId",
